update route createBulkDB
